+++ CHANGELOG.md
@@ -5,6 +5,12 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.6] - 2023-05-06
+
+### Fixed
+
+- Posting without Javascript enabled in the browser didn't detect new validation data.
+
 ## [0.8.5] - 2023-05-04
 
 ### Fixed